day_6/requirements.txt
flask
botbuilder-core

aiohttp
gunicorn

